Move MergeEnvironments() in its own util package
